requirements.txt
paho-mqtt==2.0.0
python_dateutil==2.9.0
urllib3==2.2.1